split addPlayer to smart and dumb comps
